+++ src/Symfony/Component/Security/Core/User/ChainUserChecker.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Security\Core\User;
 
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+
 final class ChainUserChecker implements UserCheckerInterface
 {
     /**
@@ -27,10 +29,16 @@ final class ChainUserChecker implements UserCheckerInterface
         }
     }
 
-    public function checkPostAuth(UserInterface $user): void
+    public function checkPostAuth(UserInterface $user /*, TokenInterface $token*/): void
     {
+        $token = 1 < \func_num_args() ? func_get_arg(1) : null;
+
         foreach ($this->checkers as $checker) {
-            $checker->checkPostAuth($user);
+            if ($token instanceof TokenInterface) {
+                $checker->checkPostAuth($user, $token);
+            } else {
+                $checker->checkPostAuth($user);
+            }
         }
     }
 }